Remove the 'new_filters' feature flag
- Update Settings
- Remove source code and specs associated with the old results page
filters
