--- app/views/new_filters/_all.html.erb
@@ -1,17 +1,0 @@
-<%= form_with url: results_path, skip_enforcing_utf8: true, class: 'app-filter', method: :get do |form| %>
-  <div class="app-filter__header">
-    <h2 class="govuk-heading-m">Filters</h2>
-  </div>
-
-  <div class="app-filter__content">
-    <%= render 'new_filters/location_provider_filter', form: form %>
-    <%= render 'new_filters/subjects_filter', form: form %>
-    <%= render 'new_filters/send_filter', form: form %>
-    <%= render 'new_filters/vacancy_filter', form: form %>
-    <%= render 'new_filters/study_type_filter', form: form %>
-    <%= render 'new_filters/qualifications_filter', form: form %>
-    <%= render 'new_filters/salary_filter', form: form %>
-    <%= render 'new_filters/hidden_fields', form: form %>
-    <%= form.submit 'Apply filters', name: nil, class: 'govuk-button govuk-!-margin-bottom-1', data: { qa: 'apply-filters' } %>
-  </div>
-<% end %>
