+++ app/views/result_filters/_all.html.erb
@@ -1,0 +1,17 @@
+<%= form_with url: results_path, skip_enforcing_utf8: true, class: 'app-filter', method: :get do |form| %>
+  <div class="app-filter__header">
+    <h2 class="govuk-heading-m">Filters</h2>
+  </div>
+
+  <div class="app-filter__content">
+    <%= render 'result_filters/location_provider_filter', form: form %>
+    <%= render 'result_filters/subjects_filter', form: form %>
+    <%= render 'result_filters/send_filter', form: form %>
+    <%= render 'result_filters/vacancy_filter', form: form %>
+    <%= render 'result_filters/study_type_filter', form: form %>
+    <%= render 'result_filters/qualifications_filter', form: form %>
+    <%= render 'result_filters/salary_filter', form: form %>
+    <%= render 'result_filters/hidden_fields', form: form %>
+    <%= form.submit 'Apply filters', name: nil, class: 'govuk-button govuk-!-margin-bottom-1', data: { qa: 'apply-filters' } %>
+  </div>
+<% end %>
