+++ app/views/result_filters/_hidden_fields.html.erb
