+++ app/views/result_filters/_location_provider_filter.html.erb
