+++ app/views/result_filters/_salary_filter.html.erb
