+++ app/views/result_filters/_send_filter.html.erb
