+++ app/views/result_filters/_study_type_filter.html.erb
