+++ app/views/result_filters/_vacancy_filter.html.erb
