+++ app/views/results/index.html.erb
@@ -9,7 +9,7 @@
 
 <div class="app-filter-layout">
   <div class="app-filter-layout__filter">
-    <%= render partial: 'new_filters/all' %>
+    <%= render partial: 'result_filters/all' %>
   </div>
   <div class="app-filter-layout__content">
     <% if @results_view.no_results_found? %>
