--- spec/features/result_filters/funding_spec.rb
@@ -1,151 +1,0 @@
-require 'rails_helper'
-
-describe 'Funding filter', type: :feature do
-  include StubbedRequests::Courses
-  include StubbedRequests::Subjects
-
-  let(:filter_page) { PageObjects::Page::ResultFilters::Funding.new }
-  let(:results_page) { PageObjects::Page::Results.new }
-
-  let(:salary_course_param_value_from_c_sharp) { '8' }
-  let(:all_course_param_value_from_c_sharp) { '15' }
-  let(:base_parameters) { results_page_parameters }
-
-  before do
-    stub_subjects
-  end
-
-  describe 'Funding filter page' do
-    before { filter_page.load }
-
-    it 'has the correct title and heading' do
-      expect(page.title).to have_content('Find courses that pay a salary')
-      expect(page).to have_content('Find courses that pay a salary')
-    end
-
-    describe 'back link' do
-      before do
-        stub_courses(query: base_parameters, course_count: 10)
-      end
-
-      it 'navigates back to the results page' do
-        filter_page.load(query: { test: 'params' })
-        filter_page.back_link.click
-
-        expect_page_to_be_displayed_with_query(
-          page: results_page,
-          expected_query_params: {
-            'fulltime' => 'false',
-            'hasvacancies' => 'true',
-            'parttime' => 'false',
-            'qualifications' => %w[QtsOnly PgdePgceWithQts Other],
-            'senCourses' => 'false',
-            'test' => 'params',
-          },
-        )
-      end
-    end
-
-    describe 'pre populated value' do
-      context 'with no salary_filter parameter' do
-        it 'the all_courses button is selected by default' do
-          expect(filter_page.all_courses).to be_checked
-          expect(filter_page.salary_courses).not_to be_checked
-        end
-      end
-
-      context 'with the salary filter set to salary courses only' do
-        before { filter_page.load(query: { 'funding' => salary_course_param_value_from_c_sharp }) }
-
-        it 'the salary_course button is selected' do
-          expect(filter_page.salary_courses).to be_checked
-          expect(filter_page.all_courses).not_to be_checked
-        end
-      end
-
-      context 'with the salary filter set to all courses' do
-        before { filter_page.load(query: { 'funding' => all_course_param_value_from_c_sharp }) }
-
-        it 'the salary_course button is selected' do
-          expect(filter_page.salary_courses).not_to be_checked
-          expect(filter_page.all_courses).to be_checked
-        end
-      end
-    end
-  end
-
-  describe 'viewing results without explicitly selecting a filter' do
-    before do
-      stub_courses(query: base_parameters, course_count: 10)
-    end
-
-    it 'lists course with or without salary' do
-      results_page.load
-
-      expect(results_page.funding_filter).to have_with_or_without_salary
-      expect(results_page.courses.count).to eq(10)
-    end
-  end
-
-  describe 'applying a filter' do
-    before do
-      stub_courses(query: base_parameters, course_count: 10)
-    end
-
-    context 'selecting all courses' do
-      before do
-        stub_courses(query: base_parameters, course_count: 10)
-      end
-
-      it 'lists the courses' do
-        results_page.load
-        results_page.funding_filter.link.click
-
-        expect(filter_page.heading.text).to eq('Find courses that pay a salary')
-
-        filter_page.all_courses.click
-        filter_page.find_courses.click
-
-        expect(results_page.heading.text).to eq('Teacher training courses 10 courses found')
-        expect(results_page.funding_filter.with_or_without_salary.text).to eq('Courses with and without salary')
-
-        expect(results_page.courses.count).to eq(10)
-      end
-    end
-
-    context 'selecting salary only courses' do
-      before do
-        stub_courses(query: base_parameters.merge('filter[funding]' => 'salary'), course_count: 10)
-      end
-
-      it 'lists the courses' do
-        results_page.load
-        results_page.funding_filter.link.click
-
-        expect(filter_page.heading.text).to eq('Find courses that pay a salary')
-
-        filter_page.salary_courses.click
-        filter_page.find_courses.click
-
-        expect(results_page.heading.text).to eq('Teacher training courses 10 courses found')
-        expect(results_page.funding_filter.with_salary.text).to eq('Only courses with a salary')
-
-        expect(results_page.courses.count).to eq(10)
-      end
-    end
-  end
-
-  describe 'submitting without applying a filter' do
-    before do
-      stub_courses(query: base_parameters, course_count: 10)
-      stub_courses(query: base_parameters.merge('filter[qualifications]' => 'QtsOnly,PgdePgceWithQts,Other'), course_count: 10)
-    end
-
-    it 'lists only courses with and without salary' do
-      results_page.load
-
-      expect(results_page.funding_filter.with_or_without_salary.text).to eq('Courses with and without salary')
-      expect(results_page.courses.count).to eq(10)
-    end
-  end
-end
