+++ spec/features/result_filters/subjects_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe 'Subject filter', type: :feature do
+RSpec.feature 'Results page new subject filter' do
   include StubbedRequests::Courses
   include StubbedRequests::SubjectAreas
   include StubbedRequests::Subjects
@@ -20,72 +20,41 @@ describe 'Subject filter', type: :feature do
       stub_courses(query: base_parameters, course_count: 10)
     end
 
-    context 'with less than 4 subjects selected' do
-      before do
-        stub_courses(
-          query: base_parameters.merge('filter[subjects]' => '00,F1'),
-          course_count: 10,
-        )
-      end
-
-      it 'lists the results' do
-        results_page.load
-        results_page.subjects_filter.link.click
-
-        expect(filter_page.heading.text).to eq(I18n.t('page_titles.subjects_filter'))
-        filter_page.subject_areas.first.subjects[0].checkbox.click
-        filter_page.subject_areas.second.subjects[3].checkbox.click
-
-        filter_page.continue.click
-
-        expect(results_page.heading.text).to eq('Teacher training courses 10 courses found')
-        expect(results_page.subjects_filter.subjects.map(&:text))
-          .to eq(
-            %w[
-              Chemistry
-              Primary
-            ],
-          )
-        expect(results_page.subjects_filter).not_to have_extra_subjects
-
-        expect(results_page.courses.count).to eq(10)
-      end
-    end
-
     context 'with subjects selected' do
       before do
         stub_courses(
           query: base_parameters.merge('filter[subjects]' => '00,01,F1,Q8,P3'),
           course_count: 10,
         )
-      end
 
-      it 'lists the results' do
         results_page.load
         results_page.subjects_filter.link.click
-
-        expect(filter_page.heading.text).to eq(I18n.t('page_titles.subjects_filter'))
         filter_page.subject_areas.first.subjects[0].checkbox.click
         filter_page.subject_areas.first.subjects[1].checkbox.click
         filter_page.subject_areas.second.subjects[3].checkbox.click
         filter_page.subject_areas.second.subjects[5].checkbox.click
         filter_page.subject_areas.second.subjects[6].checkbox.click
-
         filter_page.continue.click
+      end
 
+      it 'lists the results' do
         expect(results_page.heading.text).to eq('Teacher training courses 10 courses found')
-        expect(results_page.subjects_filter.subjects.map(&:text))
-          .to eq(
-            [
-              'Chemistry',
-              'Classics',
-              'Communication and media studies',
-              'Primary',
-            ],
-          )
-        expect(results_page.subjects_filter.extra_subjects.text).to eq('and 1 more...')
+        expect(results_page.subjects_filter.subjects.first.text).to eq(
+          'Chemistry, Classics, Communication and media studies, Primary, Primary with English',
+        )
+      end
 
-        expect(results_page.courses.count).to eq(10)
+      it 'retains the query parameters' do
+        expect_page_to_be_displayed_with_query(
+          page: results_page,
+          expected_query_params: {
+            'fulltime' => 'false',
+            'parttime' => 'false',
+            'hasvacancies' => 'true',
+            'qualifications' => %w[QtsOnly PgdePgceWithQts Other],
+            'subjects' => %w[31 32 3 5 47],
+          },
+        )
       end
     end
   end
@@ -97,32 +66,34 @@ describe 'Subject filter', type: :feature do
         'filter[send_courses]' => 'true',
       )
       stub_courses(query: query, course_count: 10)
-    end
 
-    it 'lists the results' do
       results_page.load
       results_page.subjects_filter.link.click
-
-      expect(filter_page.heading.text).to eq(I18n.t('page_titles.subjects_filter'))
       filter_page.subject_areas.first.subjects[0].checkbox.click
       filter_page.subject_areas.first.subjects[1].checkbox.click
       filter_page.subject_areas.second.subjects[3].checkbox.click
       filter_page.send_area.subjects.first.checkbox.click
       filter_page.continue.click
+    end
 
+    it 'lists the results' do
       expect(results_page.heading.text).to eq('Teacher training courses 10 courses found')
-      expect(results_page.subjects_filter).to have_send_courses
-      expect(results_page.subjects_filter.subjects.map(&:text))
-        .to eq(
-          [
-            'Chemistry',
-            'Primary',
-            'Primary with English',
-          ],
-        )
-      expect(results_page.subjects_filter).not_to have_extra_subjects
+      expect(results_page.subjects_filter.subjects.map.first.text).to eq('Chemistry, Primary, Primary with English')
+      expect(results_page.send_filter.checkbox.checked?).to be(true)
+    end
 
-      expect(results_page.courses.count).to eq(10)
+    it 'retains the query parameters' do
+      expect_page_to_be_displayed_with_query(
+        page: results_page,
+        expected_query_params: {
+          'fulltime' => 'false',
+          'parttime' => 'false',
+          'hasvacancies' => 'true',
+          'qualifications' => %w[QtsOnly PgdePgceWithQts Other],
+          'subjects' => %w[31 32 3],
+          'senCourses' => 'true',
+        },
+      )
     end
   end
 
@@ -292,20 +263,17 @@ describe 'Subject filter', type: :feature do
       )
 
       visit subject_path(subjects: %w[31], senCourses: 'true')
-      filter_page.subject_areas.first.subjects[0].checkbox.click # unselect
-      filter_page.subject_areas.first.subjects[1].checkbox.click # select a different one
-      filter_page.send_area.subjects.first.checkbox.click # unselect
-
+      filter_page.subject_areas.first.subjects[0].checkbox.click
+      filter_page.subject_areas.first.subjects[1].checkbox.click
+      filter_page.send_area.subjects.first.checkbox.click
       filter_page.continue.click
 
-      expect(results_page.subjects_filter).not_to have_send_courses
       expect(results_page.subjects_filter.subjects.map(&:text))
-          .to eq(
-            [
-              'Primary with English',
-            ],
-          )
-      expect(results_page.subjects_filter).not_to have_extra_subjects
+        .to eq(
+          [
+            'Primary with English',
+          ],
+        )
     end
   end
